\name{violin_templates}
\alias{violin_templates}
\title{List available violinplus plot templates}
\usage{
violin_templates()
}
\value{
A data frame with template IDs, titles, descriptions, palettes, thumbnails, and default parameter lists.
}
\description{
List available violinplus plot templates.
}
